Several editorial changes.
I would hope that none of them change the any of the meanings, although I
have taken the liberty of juggling with words here and there.

I've also moved the listings to lstlistings and moved the includes to
be XML only; this let me validate at least the full example documents, and I've
applied a few syntactic fixes to them, too.