Reformatting the table requirements.
For one, I feel the "we require a PARAM element" style feels a bit odd,
since people have to read on to discover that it's not just any old
PARAM but two very specific ones.

Also, I'm sure we need to make clear where the requirements on having
columns referencing TIMESYS and the PHOTCAL group come from.

I've dropped the COOSYS requirement; if I'd say anything about it at all,
I'd outlaw COOSYS-referencing columns unless the time series actually
is a positional time series, so as to not block those for later.