tighten up language concerning elements and attributes
Make sure that the use of the terms "element" and "attribute" is
in line with normal XML usage.  In some cases I think(?) that
the text was left over from earlier versions in which a new
FILTERSYS element had been considered, and in other cases it
was just a bit inconsistent.

Also modify phrasing of attribute/elements requirements a bit
for consistency.